refactor: 로딩 모션 및 템플릿 분리
