+++ src/js/ui/template.js
@@ -1,0 +1,58 @@
+import { store } from '../domain/store';
+import { MESSAGE, STORAGE_KEY } from '../constants';
+import { convertToKoreaLocaleDate } from '../utils/common';
+import NoResultImage from '../../assets/images/not_found.png';
+
+const skeletonTemplate = `
+  <div class="skeleton">
+    <div class="skeleton__image"></div>
+    <p class="skeleton__line"></p>
+    <p class="skeleton__line"></p>
+    <p class="skeleton__line"></p>
+  </div>
+`;
+
+const getFoundResultTemplate = items => {
+  const saveDatas = store.getLocalStorage(STORAGE_KEY) ?? [];
+  const resultTemplate = items
+    .map(item => {
+      const { publishedAt, channelId, title, thumbnails, channelTitle } =
+        item.snippet;
+      return `
+      <li class="video-item" data-video-id=${item.id.videoId}>
+        <a href="https://www.youtube.com/watch?v=${item.id.videoId}">
+          <img
+            src=${thumbnails.high.url}
+            alt="video-item-thumbnail" class="video-item__thumbnail">
+        </a>
+        <a href="https://www.youtube.com/watch?v=${item.id.videoId}">
+          <h4 class="video-item__title">${title}</h4>
+        </a>
+        <a href="https://www.youtube.com/channel/${channelId}"><p class="video-item__channel-name">${channelTitle}</p></a>
+        <p class="video-item__published-date">
+          ${convertToKoreaLocaleDate(publishedAt)}
+        </p>
+        <button class="video-item__save-button button" ${
+          saveDatas.includes(item.id.videoId) ? 'hidden' : ''
+        }>⬇ 저장</button>
+      </li>
+    `;
+    })
+    .join('');
+  return resultTemplate;
+};
+
+const notFoundTemplate = `
+  <div class="no-result">
+    <img class="no-result__image"
+      src=${NoResultImage}
+      alt="no-result-image"
+    >
+    <div class="no-result__description">
+      <p>${MESSAGE.NOT_FOUND}</p>
+      <p>${MESSAGE.OTHER_KEYWORD}</p>
+    </div>
+  <div>
+`;
+
+export { skeletonTemplate, getFoundResultTemplate, notFoundTemplate };
